v 1.2 changes
##DB -
# Merge users and manually updated mapping for the same.
# Updated views to fetch Description instead of Code for some master
table data.

##GUI -
# Aligned the Order & Vendor / Ship to sections as close as possible but
exact not possible because the sections are separate (fixed width is not
a good solution).
# Menu rendering delay has been minimized. Also the tab rendering of tab
is earlier now.
# Re-check and add missing default focus (revise some)

##App -
# PO print activity log restored
# Show description instead of code field, restore sort order for Lines
grid.
# Deploy CanDelete for the MasterStatus data management
# Revisited Dashboard rendering (it takes about 8-9 seconds to load it
subsequently). This is the payload we've due to KO as we're passing some
3000 records!

~~Known issue with local IST browsers (won't affect you guys)
Comments & Files date are being escalated by 1 day by JSON due to local
browser dependent date conversion.